feat: Move back to old `moduleAlias` registration method
It's no longer needed to use the complex way since the module needs to be
installed via `npm i -g knockoutcity-mod-loader` now.
This is due to prisma not being able to generate the client when run via
`npx knockoutcity-mod-loader` as well as `npx` creating a weird folder
structure the project can't handle.