feat!: Add `type` in `manifest.yaml` for mod type differentiation
A new field `type` is now existant in the `manifest` file. It specifies
if the mod is a `server-client` or a `server-only` mod.

The following differentiates the two:
- `server-client`: Will be able to run on the client and server
(used for adding items, ...).
Will not have access to the database and no connection attempt will be made.
- `server-only`: Will only be able to run on a server environment.
Will have access to the database as well as the Sandbox API module `database`.

Defaults to `server-client` if not specified.

BREAKING CHANGE: Server Mods are not required to set the `type` to
`server-only` in the `manifest` file.